integrated-app/requirements.txt
Django>=5.1.7
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
django-ckeditor>=6.7.0
django-phonenumber-field>=7.3.0
web3>=6.15.1
requests>=2.31.0
python-dotenv>=1.0.1
whitenoise>=6.6.0
gunicorn>=21.2.0 